Changed vendorScore to double to fit backend

+++ app/src/main/java/com/example/guided_project_ss21_a02_kauf_lokal_20_frontend/model/Vendor.kt
@@ -7,7 +7,7 @@ data class Vendor(
     var id: UUID,
     var name: String,
     var offerAmount: Int?,
-    var vendorScore: Int?,
+    var vendorScore: Double?,
     var logo: String,
     var websiteURL: String,
     var color: String?,
